+++ packages/orq-rc/src/models/operations/listdatasources.ts
@@ -224,7 +224,7 @@ export const ListDatasourcesData$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01KB0FSWP3WYQ3NFYH46P1G72X"),
+  _id: z.string().default("01KB2B96M8DQW2V4ARYS53KCK2"),
   display_name: z.string(),
   description: z.string().optional(),
   status: ListDatasourcesStatus$inboundSchema,
@@ -267,7 +267,7 @@ export const ListDatasourcesData$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   ListDatasourcesData
 > = z.object({
-  id: z.string().default("01KB0FSWP3WYQ3NFYH46P1G72X"),
+  id: z.string().default("01KB2B96M8DQW2V4ARYS53KCK2"),
   displayName: z.string(),
   description: z.string().optional(),
   status: ListDatasourcesStatus$outboundSchema,