+++ packages/orq-rc/RELEASES.md
@@ -5429,3 +5429,13 @@ Based on:
 - [typescript v4.1.0-rc.11] packages/orq-rc
 ### Releases
 - [NPM v4.1.0-rc.11] https://www.npmjs.com/package/@orq-ai/node/v/4.1.0-rc.11 - packages/orq-rc
+
+## 2025-11-27 09:47:39
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.661.0 (2.762.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.1.0-rc.13] packages/orq-rc
+### Releases
+- [NPM v4.1.0-rc.13] https://www.npmjs.com/package/@orq-ai/node/v/4.1.0-rc.13 - packages/orq-rc